Struct aws_sdk_pinpoint::client::fluent_builders::GetJourneyDateRangeKpi

pub struct GetJourneyDateRangeKpi<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> GetJourneyDateRangeKpi<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<GetJourneyDateRangeKpiOutput, SdkError<GetJourneyDateRangeKpiError>> where
R::Policy: SmithyRetryPolicy<GetJourneyDateRangeKpiInputOperationOutputAlias, GetJourneyDateRangeKpiOutput, GetJourneyDateRangeKpiError, GetJourneyDateRangeKpiInputOperationRetryAlias>,

pub fn application_id(self, inp: impl Into<String>) -> Self

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

pub fn set_application_id(self, input: Option<String>) -> Self

pub fn end_time(self, inp: Instant) -> Self

The last date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-26T20:00:00Z for 8:00 PM UTC July 26, 2019.

pub fn set_end_time(self, input: Option<Instant>) -> Self

pub fn journey_id(self, inp: impl Into<String>) -> Self

The unique identifier for the journey.

pub fn set_journey_id(self, input: Option<String>) -> Self

pub fn kpi_name(self, inp: impl Into<String>) -> Self

The name of the metric, also referred to as a key performance indicator (KPI), to retrieve data for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. Examples are email-open-rate and successful-delivery-rate. For a list of valid values, see the Amazon Pinpoint Developer Guide.

pub fn set_kpi_name(self, input: Option<String>) -> Self

pub fn next_token(self, inp: impl Into<String>) -> Self

The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

pub fn set_next_token(self, input: Option<String>) -> Self

pub fn page_size(self, inp: impl Into<String>) -> Self

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

pub fn set_page_size(self, input: Option<String>) -> Self

pub fn start_time(self, inp: Instant) -> Self

The first date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-19T20:00:00Z for 8:00 PM UTC July 19, 2019. This value should also be fewer than 90 days from the current day.

pub fn set_start_time(self, input: Option<Instant>) -> Self