Struct aws_sdk_pinpoint::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for Pinpoint.

This client allows ergonomic access to a Pinpoint-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn create_app(&self) -> CreateApp<C, M, R>

pub fn create_campaign(&self) -> CreateCampaign<C, M, R>

pub fn create_email_template(&self) -> CreateEmailTemplate<C, M, R>

pub fn create_export_job(&self) -> CreateExportJob<C, M, R>

pub fn create_import_job(&self) -> CreateImportJob<C, M, R>

pub fn create_journey(&self) -> CreateJourney<C, M, R>

pub fn create_push_template(&self) -> CreatePushTemplate<C, M, R>

pub fn create_recommender_configuration(
&self
) -> CreateRecommenderConfiguration<C, M, R>

pub fn create_segment(&self) -> CreateSegment<C, M, R>

pub fn create_sms_template(&self) -> CreateSmsTemplate<C, M, R>

pub fn create_voice_template(&self) -> CreateVoiceTemplate<C, M, R>

pub fn delete_adm_channel(&self) -> DeleteAdmChannel<C, M, R>

pub fn delete_apns_channel(&self) -> DeleteApnsChannel<C, M, R>

pub fn delete_apns_sandbox_channel(&self) -> DeleteApnsSandboxChannel<C, M, R>

pub fn delete_apns_voip_channel(&self) -> DeleteApnsVoipChannel<C, M, R>

pub fn delete_apns_voip_sandbox_channel(
&self
) -> DeleteApnsVoipSandboxChannel<C, M, R>

pub fn delete_app(&self) -> DeleteApp<C, M, R>

pub fn delete_baidu_channel(&self) -> DeleteBaiduChannel<C, M, R>

pub fn delete_campaign(&self) -> DeleteCampaign<C, M, R>

pub fn delete_email_channel(&self) -> DeleteEmailChannel<C, M, R>

pub fn delete_email_template(&self) -> DeleteEmailTemplate<C, M, R>

pub fn delete_endpoint(&self) -> DeleteEndpoint<C, M, R>

pub fn delete_event_stream(&self) -> DeleteEventStream<C, M, R>

pub fn delete_gcm_channel(&self) -> DeleteGcmChannel<C, M, R>

pub fn delete_journey(&self) -> DeleteJourney<C, M, R>

pub fn delete_push_template(&self) -> DeletePushTemplate<C, M, R>

pub fn delete_recommender_configuration(
&self
) -> DeleteRecommenderConfiguration<C, M, R>

pub fn delete_segment(&self) -> DeleteSegment<C, M, R>

pub fn delete_sms_channel(&self) -> DeleteSmsChannel<C, M, R>

pub fn delete_sms_template(&self) -> DeleteSmsTemplate<C, M, R>

pub fn delete_user_endpoints(&self) -> DeleteUserEndpoints<C, M, R>

pub fn delete_voice_channel(&self) -> DeleteVoiceChannel<C, M, R>

pub fn delete_voice_template(&self) -> DeleteVoiceTemplate<C, M, R>

pub fn get_adm_channel(&self) -> GetAdmChannel<C, M, R>

pub fn get_apns_channel(&self) -> GetApnsChannel<C, M, R>

pub fn get_apns_sandbox_channel(&self) -> GetApnsSandboxChannel<C, M, R>

pub fn get_apns_voip_channel(&self) -> GetApnsVoipChannel<C, M, R>

pub fn get_apns_voip_sandbox_channel(
&self
) -> GetApnsVoipSandboxChannel<C, M, R>

pub fn get_app(&self) -> GetApp<C, M, R>

pub fn get_application_date_range_kpi(
&self
) -> GetApplicationDateRangeKpi<C, M, R>

pub fn get_application_settings(&self) -> GetApplicationSettings<C, M, R>

pub fn get_apps(&self) -> GetApps<C, M, R>

pub fn get_baidu_channel(&self) -> GetBaiduChannel<C, M, R>

pub fn get_campaign(&self) -> GetCampaign<C, M, R>

pub fn get_campaign_activities(&self) -> GetCampaignActivities<C, M, R>

pub fn get_campaign_date_range_kpi(&self) -> GetCampaignDateRangeKpi<C, M, R>

pub fn get_campaigns(&self) -> GetCampaigns<C, M, R>

pub fn get_campaign_version(&self) -> GetCampaignVersion<C, M, R>

pub fn get_campaign_versions(&self) -> GetCampaignVersions<C, M, R>

pub fn get_channels(&self) -> GetChannels<C, M, R>

pub fn get_email_channel(&self) -> GetEmailChannel<C, M, R>

pub fn get_email_template(&self) -> GetEmailTemplate<C, M, R>

pub fn get_endpoint(&self) -> GetEndpoint<C, M, R>

pub fn get_event_stream(&self) -> GetEventStream<C, M, R>

pub fn get_export_job(&self) -> GetExportJob<C, M, R>

pub fn get_export_jobs(&self) -> GetExportJobs<C, M, R>

pub fn get_gcm_channel(&self) -> GetGcmChannel<C, M, R>

pub fn get_import_job(&self) -> GetImportJob<C, M, R>

pub fn get_import_jobs(&self) -> GetImportJobs<C, M, R>

pub fn get_journey(&self) -> GetJourney<C, M, R>

pub fn get_journey_date_range_kpi(&self) -> GetJourneyDateRangeKpi<C, M, R>

pub fn get_journey_execution_activity_metrics(
&self
) -> GetJourneyExecutionActivityMetrics<C, M, R>

pub fn get_journey_execution_metrics(
&self
) -> GetJourneyExecutionMetrics<C, M, R>

pub fn get_push_template(&self) -> GetPushTemplate<C, M, R>

pub fn get_recommender_configuration(
&self
) -> GetRecommenderConfiguration<C, M, R>

pub fn get_recommender_configurations(
&self
) -> GetRecommenderConfigurations<C, M, R>

pub fn get_segment(&self) -> GetSegment<C, M, R>

pub fn get_segment_export_jobs(&self) -> GetSegmentExportJobs<C, M, R>

pub fn get_segment_import_jobs(&self) -> GetSegmentImportJobs<C, M, R>

pub fn get_segments(&self) -> GetSegments<C, M, R>

pub fn get_segment_version(&self) -> GetSegmentVersion<C, M, R>

pub fn get_segment_versions(&self) -> GetSegmentVersions<C, M, R>

pub fn get_sms_channel(&self) -> GetSmsChannel<C, M, R>

pub fn get_sms_template(&self) -> GetSmsTemplate<C, M, R>

pub fn get_user_endpoints(&self) -> GetUserEndpoints<C, M, R>

pub fn get_voice_channel(&self) -> GetVoiceChannel<C, M, R>

pub fn get_voice_template(&self) -> GetVoiceTemplate<C, M, R>

pub fn list_journeys(&self) -> ListJourneys<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_templates(&self) -> ListTemplates<C, M, R>

pub fn list_template_versions(&self) -> ListTemplateVersions<C, M, R>

pub fn phone_number_validate(&self) -> PhoneNumberValidate<C, M, R>

pub fn put_events(&self) -> PutEvents<C, M, R>

pub fn put_event_stream(&self) -> PutEventStream<C, M, R>

pub fn remove_attributes(&self) -> RemoveAttributes<C, M, R>

pub fn send_messages(&self) -> SendMessages<C, M, R>

pub fn send_users_messages(&self) -> SendUsersMessages<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_adm_channel(&self) -> UpdateAdmChannel<C, M, R>

pub fn update_apns_channel(&self) -> UpdateApnsChannel<C, M, R>

pub fn update_apns_sandbox_channel(&self) -> UpdateApnsSandboxChannel<C, M, R>

pub fn update_apns_voip_channel(&self) -> UpdateApnsVoipChannel<C, M, R>

pub fn update_apns_voip_sandbox_channel(
&self
) -> UpdateApnsVoipSandboxChannel<C, M, R>

pub fn update_application_settings(&self) -> UpdateApplicationSettings<C, M, R>

pub fn update_baidu_channel(&self) -> UpdateBaiduChannel<C, M, R>

pub fn update_campaign(&self) -> UpdateCampaign<C, M, R>

pub fn update_email_channel(&self) -> UpdateEmailChannel<C, M, R>

pub fn update_email_template(&self) -> UpdateEmailTemplate<C, M, R>

pub fn update_endpoint(&self) -> UpdateEndpoint<C, M, R>

pub fn update_endpoints_batch(&self) -> UpdateEndpointsBatch<C, M, R>

pub fn update_gcm_channel(&self) -> UpdateGcmChannel<C, M, R>

pub fn update_journey(&self) -> UpdateJourney<C, M, R>

pub fn update_journey_state(&self) -> UpdateJourneyState<C, M, R>

pub fn update_push_template(&self) -> UpdatePushTemplate<C, M, R>

pub fn update_recommender_configuration(
&self
) -> UpdateRecommenderConfiguration<C, M, R>

pub fn update_segment(&self) -> UpdateSegment<C, M, R>

pub fn update_sms_channel(&self) -> UpdateSmsChannel<C, M, R>

pub fn update_sms_template(&self) -> UpdateSmsTemplate<C, M, R>

pub fn update_template_active_version(
&self
) -> UpdateTemplateActiveVersion<C, M, R>

pub fn update_voice_channel(&self) -> UpdateVoiceChannel<C, M, R>

pub fn update_voice_template(&self) -> UpdateVoiceTemplate<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self