Module aws_sdk_pinpoint::model

Modules

activities_response

See ActivitiesResponse

activity

See Activity

activity_response

See ActivityResponse

address_configuration

See AddressConfiguration

adm_channel_request

See AdmChannelRequest

adm_channel_response

See AdmChannelResponse

adm_message

See AdmMessage

android_push_notification_template

See AndroidPushNotificationTemplate

apns_channel_request

See ApnsChannelRequest

apns_channel_response

See ApnsChannelResponse

apns_message

See ApnsMessage

apns_push_notification_template

See ApnsPushNotificationTemplate

apns_sandbox_channel_request

See ApnsSandboxChannelRequest

apns_sandbox_channel_response

See ApnsSandboxChannelResponse

apns_voip_channel_request

See ApnsVoipChannelRequest

apns_voip_channel_response

See ApnsVoipChannelResponse

apns_voip_sandbox_channel_request

See ApnsVoipSandboxChannelRequest

apns_voip_sandbox_channel_response

See ApnsVoipSandboxChannelResponse

application_date_range_kpi_response

See ApplicationDateRangeKpiResponse

application_response

See ApplicationResponse

application_settings_resource

See ApplicationSettingsResource

applications_response

See ApplicationsResponse

attribute_dimension

See AttributeDimension

attributes_resource

See AttributesResource

baidu_channel_request

See BaiduChannelRequest

baidu_channel_response

See BaiduChannelResponse

baidu_message

See BaiduMessage

base_kpi_result

See BaseKpiResult

campaign_custom_message

See CampaignCustomMessage

campaign_date_range_kpi_response

See CampaignDateRangeKpiResponse

campaign_email_message

See CampaignEmailMessage

campaign_event_filter

See CampaignEventFilter

campaign_hook

See CampaignHook

campaign_limits

See CampaignLimits

campaign_response

See CampaignResponse

campaign_sms_message

See CampaignSmsMessage

campaign_state

See CampaignState

campaigns_response

See CampaignsResponse

channel_response

See ChannelResponse

channels_response

See ChannelsResponse

condition

See Condition

conditional_split_activity

See ConditionalSplitActivity

create_application_request

See CreateApplicationRequest

create_recommender_configuration_shape

See CreateRecommenderConfigurationShape

create_template_message_body

See CreateTemplateMessageBody

custom_delivery_configuration

See CustomDeliveryConfiguration

custom_message_activity

See CustomMessageActivity

default_message

See DefaultMessage

default_push_notification_message

See DefaultPushNotificationMessage

default_push_notification_template

See DefaultPushNotificationTemplate

direct_message_configuration

See DirectMessageConfiguration

email_channel_request

See EmailChannelRequest

email_channel_response

See EmailChannelResponse

email_message

See EmailMessage

email_message_activity

See EmailMessageActivity

email_template_request

See EmailTemplateRequest

email_template_response

See EmailTemplateResponse

endpoint_batch_item

See EndpointBatchItem

endpoint_batch_request

See EndpointBatchRequest

endpoint_demographic

See EndpointDemographic

endpoint_item_response

See EndpointItemResponse

endpoint_location

See EndpointLocation

endpoint_message_result

See EndpointMessageResult

endpoint_request

See EndpointRequest

endpoint_response

See EndpointResponse

endpoint_send_configuration

See EndpointSendConfiguration

endpoint_user

See EndpointUser

endpoints_response

See EndpointsResponse

event

See Event

event_condition

See EventCondition

event_dimensions

See EventDimensions

event_filter

See EventFilter

event_item_response

See EventItemResponse

event_start_condition

See EventStartCondition

event_stream

See EventStream

events_batch

See EventsBatch

events_request

See EventsRequest

events_response

See EventsResponse

export_job_request

See ExportJobRequest

export_job_resource

See ExportJobResource

export_job_response

See ExportJobResponse

export_jobs_response

See ExportJobsResponse

gcm_channel_request

See GcmChannelRequest

gcm_channel_response

See GcmChannelResponse

gcm_message

See GcmMessage

gps_coordinates

See GpsCoordinates

gps_point_dimension

See GpsPointDimension

holdout_activity

See HoldoutActivity

import_job_request

See ImportJobRequest

import_job_resource

See ImportJobResource

import_job_response

See ImportJobResponse

import_jobs_response

See ImportJobsResponse

item_response

See ItemResponse

journey_custom_message

See JourneyCustomMessage

journey_date_range_kpi_response

See JourneyDateRangeKpiResponse

journey_email_message

See JourneyEmailMessage

journey_execution_activity_metrics_response

See JourneyExecutionActivityMetricsResponse

journey_execution_metrics_response

See JourneyExecutionMetricsResponse

journey_limits

See JourneyLimits

journey_push_message

See JourneyPushMessage

journey_response

See JourneyResponse

journey_schedule

See JourneySchedule

journey_sms_message

See JourneySmsMessage

journey_state_request

See JourneyStateRequest

journeys_response

See JourneysResponse

list_recommender_configurations_response

See ListRecommenderConfigurationsResponse

message

See Message

message_body

See MessageBody

message_configuration

See MessageConfiguration

message_request

See MessageRequest

message_response

See MessageResponse

message_result

See MessageResult

metric_dimension

See MetricDimension

multi_conditional_branch

See MultiConditionalBranch

multi_conditional_split_activity

See MultiConditionalSplitActivity

number_validate_request

See NumberValidateRequest

number_validate_response

See NumberValidateResponse

public_endpoint

See PublicEndpoint

push_message_activity

See PushMessageActivity

push_notification_template_request

See PushNotificationTemplateRequest

push_notification_template_response

See PushNotificationTemplateResponse

quiet_time

See QuietTime

random_split_activity

See RandomSplitActivity

random_split_entry

See RandomSplitEntry

raw_email

See RawEmail

recency_dimension

See RecencyDimension

recommender_configuration_response

See RecommenderConfigurationResponse

result_row

See ResultRow

result_row_value

See ResultRowValue

schedule

See Schedule

segment_behaviors

See SegmentBehaviors

segment_condition

See SegmentCondition

segment_demographics

See SegmentDemographics

segment_dimensions

See SegmentDimensions

segment_group

See SegmentGroup

segment_group_list

See SegmentGroupList

segment_import_resource

See SegmentImportResource

segment_location

See SegmentLocation

segment_reference

See SegmentReference

segment_response

See SegmentResponse

segments_response

See SegmentsResponse

send_users_message_request

See SendUsersMessageRequest

send_users_message_response

See SendUsersMessageResponse

session

See Session

set_dimension

See SetDimension

simple_condition

See SimpleCondition

simple_email

See SimpleEmail

simple_email_part

See SimpleEmailPart

sms_channel_request

See SmsChannelRequest

sms_channel_response

See SmsChannelResponse

sms_message

See SmsMessage

sms_message_activity

See SmsMessageActivity

sms_template_request

See SmsTemplateRequest

sms_template_response

See SmsTemplateResponse

start_condition

See StartCondition

tags_model

See TagsModel

template

See Template

template_active_version_request

See TemplateActiveVersionRequest

template_configuration

See TemplateConfiguration

template_response

See TemplateResponse

template_version_response

See TemplateVersionResponse

template_versions_response

See TemplateVersionsResponse

templates_response

See TemplatesResponse

treatment_resource

See TreatmentResource

update_attributes_request

See UpdateAttributesRequest

update_recommender_configuration_shape

See UpdateRecommenderConfigurationShape

voice_channel_request

See VoiceChannelRequest

voice_channel_response

See VoiceChannelResponse

voice_message

See VoiceMessage

voice_template_request

See VoiceTemplateRequest

voice_template_response

See VoiceTemplateResponse

wait_activity

See WaitActivity

wait_time

See WaitTime

write_application_settings_request

See WriteApplicationSettingsRequest

write_campaign_request

See WriteCampaignRequest

write_event_stream

See WriteEventStream

write_journey_request

See WriteJourneyRequest

write_segment_request

See WriteSegmentRequest

write_treatment_resource

See WriteTreatmentResource

Structs

ActivitiesResponse

Provides information about the activities that were performed by a campaign.

Activity

Specifies the configuration and other settings for an activity in a journey.

ActivityResponse

Provides information about an activity that was performed by a campaign.

AddressConfiguration

Specifies address-based configuration settings for a message that's sent directly to an endpoint.

AdmChannelRequest

Specifies the status and settings of the ADM (Amazon Device Messaging) channel for an application.

AdmChannelResponse

Provides information about the status and settings of the ADM (Amazon Device Messaging) channel for an application.

AdmMessage

Specifies the settings for a one-time message that's sent directly to an endpoint through the ADM (Amazon Device Messaging) channel.

AndroidPushNotificationTemplate

Specifies channel-specific content and settings for a message template that can be used in push notifications that are sent through the ADM (Amazon Device Messaging), Baidu (Baidu Cloud Push), or GCM (Firebase Cloud Messaging, formerly Google Cloud Messaging) channel.

ApnsChannelRequest

Specifies the status and settings of the APNs (Apple Push Notification service) channel for an application.

ApnsChannelResponse

Provides information about the status and settings of the APNs (Apple Push Notification service) channel for an application.

ApnsMessage

Specifies the settings for a one-time message that's sent directly to an endpoint through the APNs (Apple Push Notification service) channel.

ApnsPushNotificationTemplate

Specifies channel-specific content and settings for a message template that can be used in push notifications that are sent through the APNs (Apple Push Notification service) channel.

ApnsSandboxChannelRequest

Specifies the status and settings of the APNs (Apple Push Notification service) sandbox channel for an application.

ApnsSandboxChannelResponse

Provides information about the status and settings of the APNs (Apple Push Notification service) sandbox channel for an application.

ApnsVoipChannelRequest

Specifies the status and settings of the APNs (Apple Push Notification service) VoIP channel for an application.

ApnsVoipChannelResponse

Provides information about the status and settings of the APNs (Apple Push Notification service) VoIP channel for an application.

ApnsVoipSandboxChannelRequest

Specifies the status and settings of the APNs (Apple Push Notification service) VoIP sandbox channel for an application.

ApnsVoipSandboxChannelResponse

Provides information about the status and settings of the APNs (Apple Push Notification service) VoIP sandbox channel for an application.

ApplicationDateRangeKpiResponse

Provides the results of a query that retrieved the data for a standard metric that applies to an application, and provides information about that query.

ApplicationResponse

Provides information about an application.

ApplicationSettingsResource

Provides information about an application, including the default settings for an application.

ApplicationsResponse

Provides information about all of your applications.

AttributeDimension

Specifies attribute-based criteria for including or excluding endpoints from a segment.

AttributesResource

Provides information about the type and the names of attributes that were removed from all the endpoints that are associated with an application.

BaiduChannelRequest

Specifies the status and settings of the Baidu (Baidu Cloud Push) channel for an application.

BaiduChannelResponse

Provides information about the status and settings of the Baidu (Baidu Cloud Push) channel for an application.

BaiduMessage

Specifies the settings for a one-time message that's sent directly to an endpoint through the Baidu (Baidu Cloud Push) channel.

BaseKpiResult

Provides the results of a query that retrieved the data for a standard metric that applies to an application, campaign, or journey.

CampaignCustomMessage

Specifies the contents of a message that's sent through a custom channel to recipients of a campaign.

CampaignDateRangeKpiResponse

Provides the results of a query that retrieved the data for a standard metric that applies to a campaign, and provides information about that query.

CampaignEmailMessage

Specifies the content and "From" address for an email message that's sent to recipients of a campaign.

CampaignEventFilter

Specifies the settings for events that cause a campaign to be sent.

CampaignHook

Specifies settings for invoking an AWS Lambda function that customizes a segment for a campaign.

CampaignLimits

For a campaign, specifies limits on the messages that the campaign can send. For an application, specifies the default limits for messages that campaigns in the application can send.

CampaignResponse

Provides information about the status, configuration, and other settings for a campaign.

CampaignSmsMessage

Specifies the content and settings for an SMS message that's sent to recipients of a campaign.

CampaignState

Provides information about the status of a campaign.

CampaignsResponse

Provides information about the configuration and other settings for all the campaigns that are associated with an application.

ChannelResponse

Provides information about the general settings and status of a channel for an application.

ChannelsResponse

Provides information about the general settings and status of all channels for an application, including channels that aren't enabled for the application.

Condition

Specifies the conditions to evaluate for an activity in a journey, and how to evaluate those conditions.

ConditionalSplitActivity

Specifies the settings for a yes/no split activity in a journey. This type of activity sends participants down one of two paths in a journey, based on conditions that you specify.

To create yes/no split activities that send participants down different paths based on push notification events (such as Open or Received events), your mobile app has to specify the User ID and Endpoint ID values. For more information, see Integrating Amazon Pinpoint with your application in the Amazon Pinpoint Developer Guide.

CreateApplicationRequest

Specifies the display name of an application and the tags to associate with the application.

CreateRecommenderConfigurationShape

Specifies Amazon Pinpoint configuration settings for retrieving and processing recommendation data from a recommender model.

CreateTemplateMessageBody

Provides information about a request to create a message template.

CustomDeliveryConfiguration

Specifies the delivery configuration settings for sending a campaign or campaign treatment through a custom channel. This object is required if you use the CampaignCustomMessage object to define the message to send for the campaign or campaign treatment.

CustomMessageActivity

The settings for a custom message activity. This type of activity calls an AWS Lambda function or web hook that sends messages to participants.

DefaultMessage

Specifies the default message for all channels.

DefaultPushNotificationMessage

Specifies the default settings and content for a push notification that's sent directly to an endpoint.

DefaultPushNotificationTemplate

Specifies the default settings and content for a message template that can be used in messages that are sent through a push notification channel.

DirectMessageConfiguration

Specifies the settings and content for the default message and any default messages that you tailored for specific channels.

EmailChannelRequest

Specifies the status and settings of the email channel for an application.

EmailChannelResponse

Provides information about the status and settings of the email channel for an application.

EmailMessage

Specifies the default settings and content for a one-time email message that's sent directly to an endpoint.

EmailMessageActivity

Specifies the settings for an email activity in a journey. This type of activity sends an email message to participants.

EmailTemplateRequest

Specifies the content and settings for a message template that can be used in messages that are sent through the email channel.

EmailTemplateResponse

Provides information about the content and settings for a message template that can be used in messages that are sent through the email channel.

EndpointBatchItem

Specifies an endpoint to create or update and the settings and attributes to set or change for the endpoint.

EndpointBatchRequest

Specifies a batch of endpoints to create or update and the settings and attributes to set or change for each endpoint.

EndpointDemographic

Specifies demographic information about an endpoint, such as the applicable time zone and platform.

EndpointItemResponse

Provides the status code and message that result from processing data for an endpoint.

EndpointLocation

Specifies geographic information about an endpoint.

EndpointMessageResult

Provides information about the delivery status and results of sending a message directly to an endpoint.

EndpointRequest

Specifies the channel type and other settings for an endpoint.

EndpointResponse

Provides information about the channel type and other settings for an endpoint.

EndpointSendConfiguration

Specifies the content, including message variables and attributes, to use in a message that's sent directly to an endpoint.

EndpointUser

Specifies data for one or more attributes that describe the user who's associated with an endpoint.

EndpointsResponse

Provides information about all the endpoints that are associated with a user ID.

Event

Specifies information about an event that reports data to Amazon Pinpoint.

EventCondition

Specifies the conditions to evaluate for an event that applies to an activity in a journey.

EventDimensions

Specifies the dimensions for an event filter that determines when a campaign is sent or a journey activity is performed.

EventFilter

Specifies the settings for an event that causes a campaign to be sent or a journey activity to be performed.

EventItemResponse

Provides the status code and message that result from processing an event.

EventStartCondition

Specifies the settings for an event that causes a journey activity to start.

EventStream

Specifies settings for publishing event data to an Amazon Kinesis data stream or an Amazon Kinesis Data Firehose delivery stream.

EventsBatch

Specifies a batch of endpoints and events to process.

EventsRequest

Specifies a batch of events to process.

EventsResponse

Provides information about endpoints and the events that they're associated with.

ExportJobRequest

Specifies the settings for a job that exports endpoint definitions to an Amazon Simple Storage Service (Amazon S3) bucket.

ExportJobResource

Provides information about the resource settings for a job that exports endpoint definitions to a file. The file can be added directly to an Amazon Simple Storage Service (Amazon S3) bucket by using the Amazon Pinpoint API or downloaded directly to a computer by using the Amazon Pinpoint console.

ExportJobResponse

Provides information about the status and settings of a job that exports endpoint definitions to a file. The file can be added directly to an Amazon Simple Storage Service (Amazon S3) bucket by using the Amazon Pinpoint API or downloaded directly to a computer by using the Amazon Pinpoint console.

ExportJobsResponse

Provides information about all the export jobs that are associated with an application or segment. An export job is a job that exports endpoint definitions to a file.

GcmChannelRequest

Specifies the status and settings of the GCM channel for an application. This channel enables Amazon Pinpoint to send push notifications through the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service.

GcmChannelResponse

Provides information about the status and settings of the GCM channel for an application. The GCM channel enables Amazon Pinpoint to send push notifications through the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service.

GcmMessage

Specifies the settings for a one-time message that's sent directly to an endpoint through the GCM channel. The GCM channel enables Amazon Pinpoint to send messages to the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service.

GpsCoordinates

Specifies the GPS coordinates of a location.

GpsPointDimension

Specifies GPS-based criteria for including or excluding endpoints from a segment.

HoldoutActivity

Specifies the settings for a holdout activity in a journey. This type of activity stops a journey for a specified percentage of participants.

ImportJobRequest

Specifies the settings for a job that imports endpoint definitions from an Amazon Simple Storage Service (Amazon S3) bucket.

ImportJobResource

Provides information about the resource settings for a job that imports endpoint definitions from one or more files. The files can be stored in an Amazon Simple Storage Service (Amazon S3) bucket or uploaded directly from a computer by using the Amazon Pinpoint console.

ImportJobResponse

Provides information about the status and settings of a job that imports endpoint definitions from one or more files. The files can be stored in an Amazon Simple Storage Service (Amazon S3) bucket or uploaded directly from a computer by using the Amazon Pinpoint console.

ImportJobsResponse

Provides information about the status and settings of all the import jobs that are associated with an application or segment. An import job is a job that imports endpoint definitions from one or more files.

ItemResponse

Provides information about the results of a request to create or update an endpoint that's associated with an event.

JourneyCustomMessage

Specifies the message content for a custom channel message that's sent to participants in a journey.

JourneyDateRangeKpiResponse

Provides the results of a query that retrieved the data for a standard engagement metric that applies to a journey, and provides information about that query.

JourneyEmailMessage

Specifies the "From" address for an email message that's sent to participants in a journey.

JourneyExecutionActivityMetricsResponse

Provides the results of a query that retrieved the data for a standard execution metric that applies to a journey activity, and provides information about that query.

JourneyExecutionMetricsResponse

Provides the results of a query that retrieved the data for a standard execution metric that applies to a journey, and provides information about that query.

JourneyLimits

Specifies limits on the messages that a journey can send and the number of times participants can enter a journey.

JourneyPushMessage

Specifies the message configuration for a push notification that's sent to participants in a journey.

JourneyResponse

Provides information about the status, configuration, and other settings for a journey.

JourneySchedule

Specifies the schedule settings for a journey.

JourneySmsMessage

Specifies the sender ID and message type for an SMS message that's sent to participants in a journey.

JourneyStateRequest

Changes the status of a journey.

JourneysResponse

Provides information about the status, configuration, and other settings for all the journeys that are associated with an application.

ListRecommenderConfigurationsResponse

Provides information about all the recommender model configurations that are associated with your Amazon Pinpoint account.

Message

Specifies the content and settings for a push notification that's sent to recipients of a campaign.

MessageBody

Provides information about an API request or response.

MessageConfiguration

Specifies the message configuration settings for a campaign.

MessageRequest

Specifies the configuration and other settings for a message.

MessageResponse

Provides information about the results of a request to send a message to an endpoint address.

MessageResult

Provides information about the results of sending a message directly to an endpoint address.

MetricDimension

Specifies metric-based criteria for including or excluding endpoints from a segment. These criteria derive from custom metrics that you define for endpoints.

MultiConditionalBranch

Specifies a condition to evaluate for an activity path in a journey.

MultiConditionalSplitActivity

Specifies the settings for a multivariate split activity in a journey. This type of activity sends participants down one of as many as five paths (including a default Else path) in a journey, based on conditions that you specify.

To create multivariate split activities that send participants down different paths based on push notification events (such as Open or Received events), your mobile app has to specify the User ID and Endpoint ID values. For more information, see Integrating Amazon Pinpoint with your application in the Amazon Pinpoint Developer Guide.

NumberValidateRequest

Specifies a phone number to validate and retrieve information about.

NumberValidateResponse

Provides information about a phone number.

PublicEndpoint

Specifies the properties and attributes of an endpoint that's associated with an event.

PushMessageActivity

Specifies the settings for a push notification activity in a journey. This type of activity sends a push notification to participants.

PushNotificationTemplateRequest

Specifies the content and settings for a message template that can be used in messages that are sent through a push notification channel.

PushNotificationTemplateResponse

Provides information about the content and settings for a message template that can be used in messages that are sent through a push notification channel.

QuietTime

Specifies the start and end times that define a time range when messages aren't sent to endpoints.

RandomSplitActivity

Specifies the settings for a random split activity in a journey. This type of activity randomly sends specified percentages of participants down one of as many as five paths in a journey, based on conditions that you specify.

RandomSplitEntry

Specifies the settings for a path in a random split activity in a journey.

RawEmail

Specifies the contents of an email message, represented as a raw MIME message.

RecencyDimension

Specifies criteria for including or excluding endpoints from a segment based on how recently an endpoint was active.

RecommenderConfigurationResponse

Provides information about Amazon Pinpoint configuration settings for retrieving and processing data from a recommender model.

ResultRow

Provides the results of a query that retrieved the data for a standard metric that applies to an application, campaign, or journey.

ResultRowValue

Provides a single value and metadata about that value as part of an array of query results for a standard metric that applies to an application, campaign, or journey.

Schedule

Specifies the schedule settings for a campaign.

SegmentBehaviors

Specifies dimension settings for including or excluding endpoints from a segment based on how recently an endpoint was active.

SegmentCondition

Specifies a segment to associate with an activity in a journey.

SegmentDemographics

Specifies demographic-based dimension settings for including or excluding endpoints from a segment. These settings derive from characteristics of endpoint devices, such as platform, make, and model.

SegmentDimensions

Specifies the dimension settings for a segment.

SegmentGroup

Specifies the base segments and dimensions for a segment, and the relationships between these base segments and dimensions.

SegmentGroupList

Specifies the settings that define the relationships between segment groups for a segment.

SegmentImportResource

Provides information about the import job that created a segment. An import job is a job that creates a user segment by importing endpoint definitions.

SegmentLocation

Specifies geographical dimension settings for a segment.

SegmentReference

Specifies the segment identifier and version of a segment.

SegmentResponse

Provides information about the configuration, dimension, and other settings for a segment.

SegmentsResponse

Provides information about all the segments that are associated with an application.

SendUsersMessageRequest

Specifies the configuration and other settings for a message to send to all the endpoints that are associated with a list of users.

SendUsersMessageResponse

Provides information about which users and endpoints a message was sent to.

Session

Provides information about a session.

SetDimension

Specifies the dimension type and values for a segment dimension.

SimpleCondition

Specifies a condition to evaluate for an activity in a journey.

SimpleEmail

Specifies the contents of an email message, composed of a subject, a text part, and an HTML part.

SimpleEmailPart

Specifies the subject or body of an email message, represented as textual email data and the applicable character set.

SmsChannelRequest

Specifies the status and settings of the SMS channel for an application.

SmsChannelResponse

Provides information about the status and settings of the SMS channel for an application.

SmsMessage

Specifies the default settings for a one-time SMS message that's sent directly to an endpoint.

SmsMessageActivity

Specifies the settings for an SMS activity in a journey. This type of activity sends a text message to participants.

SmsTemplateRequest

Specifies the content and settings for a message template that can be used in text messages that are sent through the SMS channel.

SmsTemplateResponse

Provides information about the content and settings for a message template that can be used in text messages that are sent through the SMS channel.

StartCondition

Specifies the conditions for the first activity in a journey. This activity and its conditions determine which users are participants in a journey.

TagsModel

Specifies the tags (keys and values) for an application, campaign, message template, or segment.

Template

Specifies the name and version of the message template to use for the message.

TemplateActiveVersionRequest

Specifies which version of a message template to use as the active version of the template.

TemplateConfiguration

Specifies the message template to use for the message, for each type of channel.

TemplateResponse

Provides information about a message template that's associated with your Amazon Pinpoint account.

TemplateVersionResponse

Provides information about a specific version of a message template.

TemplateVersionsResponse

Provides information about all the versions of a specific message template.

TemplatesResponse

Provides information about all the message templates that are associated with your Amazon Pinpoint account.

TreatmentResource

Specifies the settings for a campaign treatment. A treatment is a variation of a campaign that's used for A/B testing of a campaign.

UpdateAttributesRequest

Specifies one or more attributes to remove from all the endpoints that are associated with an application.

UpdateRecommenderConfigurationShape

Specifies Amazon Pinpoint configuration settings for retrieving and processing recommendation data from a recommender model.

VoiceChannelRequest

Specifies the status and settings of the voice channel for an application.

VoiceChannelResponse

Provides information about the status and settings of the voice channel for an application.

VoiceMessage

Specifies the settings for a one-time voice message that's sent directly to an endpoint through the voice channel.

VoiceTemplateRequest

Specifies the content and settings for a message template that can be used in messages that are sent through the voice channel.

VoiceTemplateResponse

Provides information about the content and settings for a message template that can be used in messages that are sent through the voice channel.

WaitActivity

Specifies the settings for a wait activity in a journey. This type of activity waits for a certain amount of time or until a specific date and time before moving participants to the next activity in a journey.

WaitTime

Specifies a duration or a date and time that indicates when Amazon Pinpoint determines whether an activity's conditions have been met or an activity moves participants to the next activity in a journey.

WriteApplicationSettingsRequest

Specifies the default settings for an application.

WriteCampaignRequest

Specifies the configuration and other settings for a campaign.

WriteEventStream

Specifies the Amazon Resource Name (ARN) of an event stream to publish events to and the AWS Identity and Access Management (IAM) role to use when publishing those events.

WriteJourneyRequest

Specifies the configuration and other settings for a journey.

WriteSegmentRequest

Specifies the configuration, dimension, and other settings for a segment. A WriteSegmentRequest object can include a Dimensions object or a SegmentGroups object, but not both.

WriteTreatmentResource

Specifies the settings for a campaign treatment. A treatment is a variation of a campaign that's used for A/B testing of a campaign.

Enums

Action

AttributeType

CampaignStatus

ChannelType

DeliveryStatus

DimensionType

Duration

EndpointTypesElement

FilterType

Format

Frequency

Include

JobStatus

MessageType

Mode

Operator

RecencyType

SegmentType

SourceType

State

TemplateType

Type