Struct aws_sdk_pinpoint::model::apns_message::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ApnsMessage

Implementations

impl Builder

pub fn apns_push_type(self, input: impl Into<String>) -> Self

The type of push notification to send. Valid values are:

alert - For a standard notification that's displayed on recipients' devices and prompts a recipient to interact with the notification.

background - For a silent notification that delivers content in the background and isn't displayed on recipients' devices.

complication - For a notification that contains update information for an app’s complication timeline.

fileprovider - For a notification that signals changes to a File Provider extension.

mdm - For a notification that tells managed devices to contact the MDM server.

voip - For a notification that provides information about an incoming VoIP call.

Amazon Pinpoint specifies this value in the apns-push-type request header when it sends the notification message to APNs. If you don't specify a value for this property, Amazon Pinpoint sets the value to alert or background automatically, based on the value that you specify for the SilentPush or RawContent property of the message.

For more information about the apns-push-type request header, see Sending Notification Requests to APNs on the Apple Developer website.

pub fn set_apns_push_type(self, input: Option<String>) -> Self

pub fn action(self, input: Action) -> Self

The action to occur if the recipient taps the push notification. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This setting uses the deep-linking features of the iOS platform.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

pub fn set_action(self, input: Option<Action>) -> Self

pub fn badge(self, input: i32) -> Self

The key that indicates whether and how to modify the badge of your app's icon when the recipient receives the push notification. If this key isn't included in the dictionary, the badge doesn't change. To remove the badge, set this value to 0.

pub fn set_badge(self, input: Option<i32>) -> Self

pub fn body(self, input: impl Into<String>) -> Self

The body of the notification message.

pub fn set_body(self, input: Option<String>) -> Self

pub fn category(self, input: impl Into<String>) -> Self

The key that indicates the notification type for the push notification. This key is a value that's defined by the identifier property of one of your app's registered categories.

pub fn set_category(self, input: Option<String>) -> Self

pub fn collapse_id(self, input: impl Into<String>) -> Self

An arbitrary identifier that, if assigned to multiple messages, APNs uses to coalesce the messages into a single push notification instead of delivering each message individually. This value can't exceed 64 bytes.

Amazon Pinpoint specifies this value in the apns-collapse-id request header when it sends the notification message to APNs.

pub fn set_collapse_id(self, input: Option<String>) -> Self

pub fn data(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_data(self, input: Option<HashMap<String, String>>) -> Self

pub fn media_url(self, input: impl Into<String>) -> Self

The URL of an image or video to display in the push notification.

pub fn set_media_url(self, input: Option<String>) -> Self

pub fn preferred_authentication_method(self, input: impl Into<String>) -> Self

The authentication method that you want Amazon Pinpoint to use when authenticating with APNs, CERTIFICATE or TOKEN.

pub fn set_preferred_authentication_method(self, input: Option<String>) -> Self

pub fn priority(self, input: impl Into<String>) -> Self

para>5 - Low priority, the notification might be delayed, delivered as part of a group, or throttled.
/listitem>

10 - High priority, the notification is sent immediately. This is the default value. A high priority notification should trigger an alert, play a sound, or badge your app's icon on the recipient's device.

/para>
Amazon Pinpoint specifies this value in the apns-priority request header when it sends the notification message to APNs.

The equivalent values for Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), are normal, for 5, and high, for 10. If you specify an FCM value for this property, Amazon Pinpoint accepts and converts the value to the corresponding APNs value.

pub fn set_priority(self, input: Option<String>) -> Self

pub fn raw_content(self, input: impl Into<String>) -> Self

The raw, JSON-formatted string to use as the payload for the notification message. If specified, this value overrides all other content for the message.

If you specify the raw content of an APNs push notification, the message payload has to include the content-available key. The value of the content-available key has to be an integer, and can only be 0 or 1. If you're sending a standard notification, set the value of content-available to 0. If you're sending a silent (background) notification, set the value of content-available to 1. Additionally, silent notification payloads can't include the alert, badge, or sound keys. For more information, see Generating a Remote Notification and Pushing Background Updates to Your App on the Apple Developer website.

pub fn set_raw_content(self, input: Option<String>) -> Self

pub fn silent_push(self, input: bool) -> Self

Specifies whether the notification is a silent push notification. A silent (or background) push notification isn't displayed on recipients' devices. You can use silent push notifications to make small updates to your app, or to display messages in an in-app message center.

Amazon Pinpoint uses this property to determine the correct value for the apns-push-type request header when it sends the notification message to APNs. If you specify a value of true for this property, Amazon Pinpoint sets the value for the apns-push-type header field to background.

If you specify the raw content of an APNs push notification, the message payload has to include the content-available key. For silent (background) notifications, set the value of content-available to 1. Additionally, the message payload for a silent notification can't include the alert, badge, or sound keys. For more information, see Generating a Remote Notification and Pushing Background Updates to Your App on the Apple Developer website.

Apple has indicated that they will throttle "excessive" background notifications based on current traffic volumes. To prevent your notifications being throttled, Apple recommends that you send no more than 3 silent push notifications to each recipient per hour.

pub fn set_silent_push(self, input: Option<bool>) -> Self

pub fn sound(self, input: impl Into<String>) -> Self

The key for the sound to play when the recipient receives the push notification. The value for this key is the name of a sound file in your app's main bundle or the Library/Sounds folder in your app's data container. If the sound file can't be found or you specify default for the value, the system plays the default alert sound.

pub fn set_sound(self, input: Option<String>) -> Self

pub fn substitutions(
self,
k: impl Into<String>,
v: impl Into<Vec<String>>
) -> Self

pub fn set_substitutions(
self,
input: Option<HashMap<String, Vec<String>>>
) -> Self

pub fn thread_id(self, input: impl Into<String>) -> Self

The key that represents your app-specific identifier for grouping notifications. If you provide a Notification Content app extension, you can use this value to group your notifications together.

pub fn set_thread_id(self, input: Option<String>) -> Self

pub fn time_to_live(self, input: i32) -> Self

The amount of time, in seconds, that APNs should store and attempt to deliver the push notification, if the service is unable to deliver the notification the first time. If this value is 0, APNs treats the notification as if it expires immediately and the service doesn't store or try to deliver the notification again.

Amazon Pinpoint specifies this value in the apns-expiration request header when it sends the notification message to APNs.

pub fn set_time_to_live(self, input: Option<i32>) -> Self

pub fn title(self, input: impl Into<String>) -> Self

The title to display above the notification message on the recipient's device.

pub fn set_title(self, input: Option<String>) -> Self

pub fn url(self, input: impl Into<String>) -> Self

The URL to open in the recipient's default mobile browser, if a recipient taps the push notification and the value of the Action property is URL.

pub fn set_url(self, input: Option<String>) -> Self

pub fn build(self) -> ApnsMessage

Consumes the builder and constructs a ApnsMessage