Struct aws_sdk_pinpoint::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for Pinpoint.

This client allows ergonomic access to a Pinpoint-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn create_app(&self) -> CreateApp<C, M, R>

Constructs a fluent builder for the CreateApp operation.

See CreateApp for more information about the operation and its arguments.

pub fn create_campaign(&self) -> CreateCampaign<C, M, R>

Constructs a fluent builder for the CreateCampaign operation.

See CreateCampaign for more information about the operation and its arguments.

pub fn create_email_template(&self) -> CreateEmailTemplate<C, M, R>

Constructs a fluent builder for the CreateEmailTemplate operation.

See CreateEmailTemplate for more information about the operation and its arguments.

pub fn create_export_job(&self) -> CreateExportJob<C, M, R>

Constructs a fluent builder for the CreateExportJob operation.

See CreateExportJob for more information about the operation and its arguments.

pub fn create_import_job(&self) -> CreateImportJob<C, M, R>

Constructs a fluent builder for the CreateImportJob operation.

See CreateImportJob for more information about the operation and its arguments.

pub fn create_in_app_template(&self) -> CreateInAppTemplate<C, M, R>

Constructs a fluent builder for the CreateInAppTemplate operation.

See CreateInAppTemplate for more information about the operation and its arguments.

pub fn create_journey(&self) -> CreateJourney<C, M, R>

Constructs a fluent builder for the CreateJourney operation.

See CreateJourney for more information about the operation and its arguments.

pub fn create_push_template(&self) -> CreatePushTemplate<C, M, R>

Constructs a fluent builder for the CreatePushTemplate operation.

See CreatePushTemplate for more information about the operation and its arguments.

pub fn create_recommender_configuration(
&self
) -> CreateRecommenderConfiguration<C, M, R>

Constructs a fluent builder for the CreateRecommenderConfiguration operation.

See CreateRecommenderConfiguration for more information about the operation and its arguments.

pub fn create_segment(&self) -> CreateSegment<C, M, R>

Constructs a fluent builder for the CreateSegment operation.

See CreateSegment for more information about the operation and its arguments.

pub fn create_sms_template(&self) -> CreateSmsTemplate<C, M, R>

Constructs a fluent builder for the CreateSmsTemplate operation.

See CreateSmsTemplate for more information about the operation and its arguments.

pub fn create_voice_template(&self) -> CreateVoiceTemplate<C, M, R>

Constructs a fluent builder for the CreateVoiceTemplate operation.

See CreateVoiceTemplate for more information about the operation and its arguments.

pub fn delete_adm_channel(&self) -> DeleteAdmChannel<C, M, R>

Constructs a fluent builder for the DeleteAdmChannel operation.

See DeleteAdmChannel for more information about the operation and its arguments.

pub fn delete_apns_channel(&self) -> DeleteApnsChannel<C, M, R>

Constructs a fluent builder for the DeleteApnsChannel operation.

See DeleteApnsChannel for more information about the operation and its arguments.

pub fn delete_apns_sandbox_channel(&self) -> DeleteApnsSandboxChannel<C, M, R>

Constructs a fluent builder for the DeleteApnsSandboxChannel operation.

See DeleteApnsSandboxChannel for more information about the operation and its arguments.

pub fn delete_apns_voip_channel(&self) -> DeleteApnsVoipChannel<C, M, R>

Constructs a fluent builder for the DeleteApnsVoipChannel operation.

See DeleteApnsVoipChannel for more information about the operation and its arguments.

pub fn delete_apns_voip_sandbox_channel(
&self
) -> DeleteApnsVoipSandboxChannel<C, M, R>

Constructs a fluent builder for the DeleteApnsVoipSandboxChannel operation.

See DeleteApnsVoipSandboxChannel for more information about the operation and its arguments.

pub fn delete_app(&self) -> DeleteApp<C, M, R>

Constructs a fluent builder for the DeleteApp operation.

See DeleteApp for more information about the operation and its arguments.

pub fn delete_baidu_channel(&self) -> DeleteBaiduChannel<C, M, R>

Constructs a fluent builder for the DeleteBaiduChannel operation.

See DeleteBaiduChannel for more information about the operation and its arguments.

pub fn delete_campaign(&self) -> DeleteCampaign<C, M, R>

Constructs a fluent builder for the DeleteCampaign operation.

See DeleteCampaign for more information about the operation and its arguments.

pub fn delete_email_channel(&self) -> DeleteEmailChannel<C, M, R>

Constructs a fluent builder for the DeleteEmailChannel operation.

See DeleteEmailChannel for more information about the operation and its arguments.

pub fn delete_email_template(&self) -> DeleteEmailTemplate<C, M, R>

Constructs a fluent builder for the DeleteEmailTemplate operation.

See DeleteEmailTemplate for more information about the operation and its arguments.

pub fn delete_endpoint(&self) -> DeleteEndpoint<C, M, R>

Constructs a fluent builder for the DeleteEndpoint operation.

See DeleteEndpoint for more information about the operation and its arguments.

pub fn delete_event_stream(&self) -> DeleteEventStream<C, M, R>

Constructs a fluent builder for the DeleteEventStream operation.

See DeleteEventStream for more information about the operation and its arguments.

pub fn delete_gcm_channel(&self) -> DeleteGcmChannel<C, M, R>

Constructs a fluent builder for the DeleteGcmChannel operation.

See DeleteGcmChannel for more information about the operation and its arguments.

pub fn delete_in_app_template(&self) -> DeleteInAppTemplate<C, M, R>

Constructs a fluent builder for the DeleteInAppTemplate operation.

See DeleteInAppTemplate for more information about the operation and its arguments.

pub fn delete_journey(&self) -> DeleteJourney<C, M, R>

Constructs a fluent builder for the DeleteJourney operation.

See DeleteJourney for more information about the operation and its arguments.

pub fn delete_push_template(&self) -> DeletePushTemplate<C, M, R>

Constructs a fluent builder for the DeletePushTemplate operation.

See DeletePushTemplate for more information about the operation and its arguments.

pub fn delete_recommender_configuration(
&self
) -> DeleteRecommenderConfiguration<C, M, R>

Constructs a fluent builder for the DeleteRecommenderConfiguration operation.

See DeleteRecommenderConfiguration for more information about the operation and its arguments.

pub fn delete_segment(&self) -> DeleteSegment<C, M, R>

Constructs a fluent builder for the DeleteSegment operation.

See DeleteSegment for more information about the operation and its arguments.

pub fn delete_sms_channel(&self) -> DeleteSmsChannel<C, M, R>

Constructs a fluent builder for the DeleteSmsChannel operation.

See DeleteSmsChannel for more information about the operation and its arguments.

pub fn delete_sms_template(&self) -> DeleteSmsTemplate<C, M, R>

Constructs a fluent builder for the DeleteSmsTemplate operation.

See DeleteSmsTemplate for more information about the operation and its arguments.

pub fn delete_user_endpoints(&self) -> DeleteUserEndpoints<C, M, R>

Constructs a fluent builder for the DeleteUserEndpoints operation.

See DeleteUserEndpoints for more information about the operation and its arguments.

pub fn delete_voice_channel(&self) -> DeleteVoiceChannel<C, M, R>

Constructs a fluent builder for the DeleteVoiceChannel operation.

See DeleteVoiceChannel for more information about the operation and its arguments.

pub fn delete_voice_template(&self) -> DeleteVoiceTemplate<C, M, R>

Constructs a fluent builder for the DeleteVoiceTemplate operation.

See DeleteVoiceTemplate for more information about the operation and its arguments.

pub fn get_adm_channel(&self) -> GetAdmChannel<C, M, R>

Constructs a fluent builder for the GetAdmChannel operation.

See GetAdmChannel for more information about the operation and its arguments.

pub fn get_apns_channel(&self) -> GetApnsChannel<C, M, R>

Constructs a fluent builder for the GetApnsChannel operation.

See GetApnsChannel for more information about the operation and its arguments.

pub fn get_apns_sandbox_channel(&self) -> GetApnsSandboxChannel<C, M, R>

Constructs a fluent builder for the GetApnsSandboxChannel operation.

See GetApnsSandboxChannel for more information about the operation and its arguments.

pub fn get_apns_voip_channel(&self) -> GetApnsVoipChannel<C, M, R>

Constructs a fluent builder for the GetApnsVoipChannel operation.

See GetApnsVoipChannel for more information about the operation and its arguments.

pub fn get_apns_voip_sandbox_channel(
&self
) -> GetApnsVoipSandboxChannel<C, M, R>

Constructs a fluent builder for the GetApnsVoipSandboxChannel operation.

See GetApnsVoipSandboxChannel for more information about the operation and its arguments.

pub fn get_app(&self) -> GetApp<C, M, R>

Constructs a fluent builder for the GetApp operation.

See GetApp for more information about the operation and its arguments.

pub fn get_application_date_range_kpi(
&self
) -> GetApplicationDateRangeKpi<C, M, R>

Constructs a fluent builder for the GetApplicationDateRangeKpi operation.

See GetApplicationDateRangeKpi for more information about the operation and its arguments.

pub fn get_application_settings(&self) -> GetApplicationSettings<C, M, R>

Constructs a fluent builder for the GetApplicationSettings operation.

See GetApplicationSettings for more information about the operation and its arguments.

pub fn get_apps(&self) -> GetApps<C, M, R>

Constructs a fluent builder for the GetApps operation.

See GetApps for more information about the operation and its arguments.

pub fn get_baidu_channel(&self) -> GetBaiduChannel<C, M, R>

Constructs a fluent builder for the GetBaiduChannel operation.

See GetBaiduChannel for more information about the operation and its arguments.

pub fn get_campaign(&self) -> GetCampaign<C, M, R>

Constructs a fluent builder for the GetCampaign operation.

See GetCampaign for more information about the operation and its arguments.

pub fn get_campaign_activities(&self) -> GetCampaignActivities<C, M, R>

Constructs a fluent builder for the GetCampaignActivities operation.

See GetCampaignActivities for more information about the operation and its arguments.

pub fn get_campaign_date_range_kpi(&self) -> GetCampaignDateRangeKpi<C, M, R>

Constructs a fluent builder for the GetCampaignDateRangeKpi operation.

See GetCampaignDateRangeKpi for more information about the operation and its arguments.

pub fn get_campaigns(&self) -> GetCampaigns<C, M, R>

Constructs a fluent builder for the GetCampaigns operation.

See GetCampaigns for more information about the operation and its arguments.

pub fn get_campaign_version(&self) -> GetCampaignVersion<C, M, R>

Constructs a fluent builder for the GetCampaignVersion operation.

See GetCampaignVersion for more information about the operation and its arguments.

pub fn get_campaign_versions(&self) -> GetCampaignVersions<C, M, R>

Constructs a fluent builder for the GetCampaignVersions operation.

See GetCampaignVersions for more information about the operation and its arguments.

pub fn get_channels(&self) -> GetChannels<C, M, R>

Constructs a fluent builder for the GetChannels operation.

See GetChannels for more information about the operation and its arguments.

pub fn get_email_channel(&self) -> GetEmailChannel<C, M, R>

Constructs a fluent builder for the GetEmailChannel operation.

See GetEmailChannel for more information about the operation and its arguments.

pub fn get_email_template(&self) -> GetEmailTemplate<C, M, R>

Constructs a fluent builder for the GetEmailTemplate operation.

See GetEmailTemplate for more information about the operation and its arguments.

pub fn get_endpoint(&self) -> GetEndpoint<C, M, R>

Constructs a fluent builder for the GetEndpoint operation.

See GetEndpoint for more information about the operation and its arguments.

pub fn get_event_stream(&self) -> GetEventStream<C, M, R>

Constructs a fluent builder for the GetEventStream operation.

See GetEventStream for more information about the operation and its arguments.

pub fn get_export_job(&self) -> GetExportJob<C, M, R>

Constructs a fluent builder for the GetExportJob operation.

See GetExportJob for more information about the operation and its arguments.

pub fn get_export_jobs(&self) -> GetExportJobs<C, M, R>

Constructs a fluent builder for the GetExportJobs operation.

See GetExportJobs for more information about the operation and its arguments.

pub fn get_gcm_channel(&self) -> GetGcmChannel<C, M, R>

Constructs a fluent builder for the GetGcmChannel operation.

See GetGcmChannel for more information about the operation and its arguments.

pub fn get_import_job(&self) -> GetImportJob<C, M, R>

Constructs a fluent builder for the GetImportJob operation.

See GetImportJob for more information about the operation and its arguments.

pub fn get_import_jobs(&self) -> GetImportJobs<C, M, R>

Constructs a fluent builder for the GetImportJobs operation.

See GetImportJobs for more information about the operation and its arguments.

pub fn get_in_app_messages(&self) -> GetInAppMessages<C, M, R>

Constructs a fluent builder for the GetInAppMessages operation.

See GetInAppMessages for more information about the operation and its arguments.

pub fn get_in_app_template(&self) -> GetInAppTemplate<C, M, R>

Constructs a fluent builder for the GetInAppTemplate operation.

See GetInAppTemplate for more information about the operation and its arguments.

pub fn get_journey(&self) -> GetJourney<C, M, R>

Constructs a fluent builder for the GetJourney operation.

See GetJourney for more information about the operation and its arguments.

pub fn get_journey_date_range_kpi(&self) -> GetJourneyDateRangeKpi<C, M, R>

Constructs a fluent builder for the GetJourneyDateRangeKpi operation.

See GetJourneyDateRangeKpi for more information about the operation and its arguments.

pub fn get_journey_execution_activity_metrics(
&self
) -> GetJourneyExecutionActivityMetrics<C, M, R>

Constructs a fluent builder for the GetJourneyExecutionActivityMetrics operation.

See GetJourneyExecutionActivityMetrics for more information about the operation and its arguments.

pub fn get_journey_execution_metrics(
&self
) -> GetJourneyExecutionMetrics<C, M, R>

Constructs a fluent builder for the GetJourneyExecutionMetrics operation.

See GetJourneyExecutionMetrics for more information about the operation and its arguments.

pub fn get_push_template(&self) -> GetPushTemplate<C, M, R>

Constructs a fluent builder for the GetPushTemplate operation.

See GetPushTemplate for more information about the operation and its arguments.

pub fn get_recommender_configuration(
&self
) -> GetRecommenderConfiguration<C, M, R>

Constructs a fluent builder for the GetRecommenderConfiguration operation.

See GetRecommenderConfiguration for more information about the operation and its arguments.

pub fn get_recommender_configurations(
&self
) -> GetRecommenderConfigurations<C, M, R>

Constructs a fluent builder for the GetRecommenderConfigurations operation.

See GetRecommenderConfigurations for more information about the operation and its arguments.

pub fn get_segment(&self) -> GetSegment<C, M, R>

Constructs a fluent builder for the GetSegment operation.

See GetSegment for more information about the operation and its arguments.

pub fn get_segment_export_jobs(&self) -> GetSegmentExportJobs<C, M, R>

Constructs a fluent builder for the GetSegmentExportJobs operation.

See GetSegmentExportJobs for more information about the operation and its arguments.

pub fn get_segment_import_jobs(&self) -> GetSegmentImportJobs<C, M, R>

Constructs a fluent builder for the GetSegmentImportJobs operation.

See GetSegmentImportJobs for more information about the operation and its arguments.

pub fn get_segments(&self) -> GetSegments<C, M, R>

Constructs a fluent builder for the GetSegments operation.

See GetSegments for more information about the operation and its arguments.

pub fn get_segment_version(&self) -> GetSegmentVersion<C, M, R>

Constructs a fluent builder for the GetSegmentVersion operation.

See GetSegmentVersion for more information about the operation and its arguments.

pub fn get_segment_versions(&self) -> GetSegmentVersions<C, M, R>

Constructs a fluent builder for the GetSegmentVersions operation.

See GetSegmentVersions for more information about the operation and its arguments.

pub fn get_sms_channel(&self) -> GetSmsChannel<C, M, R>

Constructs a fluent builder for the GetSmsChannel operation.

See GetSmsChannel for more information about the operation and its arguments.

pub fn get_sms_template(&self) -> GetSmsTemplate<C, M, R>

Constructs a fluent builder for the GetSmsTemplate operation.

See GetSmsTemplate for more information about the operation and its arguments.

pub fn get_user_endpoints(&self) -> GetUserEndpoints<C, M, R>

Constructs a fluent builder for the GetUserEndpoints operation.

See GetUserEndpoints for more information about the operation and its arguments.

pub fn get_voice_channel(&self) -> GetVoiceChannel<C, M, R>

Constructs a fluent builder for the GetVoiceChannel operation.

See GetVoiceChannel for more information about the operation and its arguments.

pub fn get_voice_template(&self) -> GetVoiceTemplate<C, M, R>

Constructs a fluent builder for the GetVoiceTemplate operation.

See GetVoiceTemplate for more information about the operation and its arguments.

pub fn list_journeys(&self) -> ListJourneys<C, M, R>

Constructs a fluent builder for the ListJourneys operation.

See ListJourneys for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn list_templates(&self) -> ListTemplates<C, M, R>

Constructs a fluent builder for the ListTemplates operation.

See ListTemplates for more information about the operation and its arguments.

pub fn list_template_versions(&self) -> ListTemplateVersions<C, M, R>

Constructs a fluent builder for the ListTemplateVersions operation.

See ListTemplateVersions for more information about the operation and its arguments.

pub fn phone_number_validate(&self) -> PhoneNumberValidate<C, M, R>

Constructs a fluent builder for the PhoneNumberValidate operation.

See PhoneNumberValidate for more information about the operation and its arguments.

pub fn put_events(&self) -> PutEvents<C, M, R>

Constructs a fluent builder for the PutEvents operation.

See PutEvents for more information about the operation and its arguments.

pub fn put_event_stream(&self) -> PutEventStream<C, M, R>

Constructs a fluent builder for the PutEventStream operation.

See PutEventStream for more information about the operation and its arguments.

pub fn remove_attributes(&self) -> RemoveAttributes<C, M, R>

Constructs a fluent builder for the RemoveAttributes operation.

See RemoveAttributes for more information about the operation and its arguments.

pub fn send_messages(&self) -> SendMessages<C, M, R>

Constructs a fluent builder for the SendMessages operation.

See SendMessages for more information about the operation and its arguments.

pub fn send_users_messages(&self) -> SendUsersMessages<C, M, R>

Constructs a fluent builder for the SendUsersMessages operation.

See SendUsersMessages for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_adm_channel(&self) -> UpdateAdmChannel<C, M, R>

Constructs a fluent builder for the UpdateAdmChannel operation.

See UpdateAdmChannel for more information about the operation and its arguments.

pub fn update_apns_channel(&self) -> UpdateApnsChannel<C, M, R>

Constructs a fluent builder for the UpdateApnsChannel operation.

See UpdateApnsChannel for more information about the operation and its arguments.

pub fn update_apns_sandbox_channel(&self) -> UpdateApnsSandboxChannel<C, M, R>

Constructs a fluent builder for the UpdateApnsSandboxChannel operation.

See UpdateApnsSandboxChannel for more information about the operation and its arguments.

pub fn update_apns_voip_channel(&self) -> UpdateApnsVoipChannel<C, M, R>

Constructs a fluent builder for the UpdateApnsVoipChannel operation.

See UpdateApnsVoipChannel for more information about the operation and its arguments.

pub fn update_apns_voip_sandbox_channel(
&self
) -> UpdateApnsVoipSandboxChannel<C, M, R>

Constructs a fluent builder for the UpdateApnsVoipSandboxChannel operation.

See UpdateApnsVoipSandboxChannel for more information about the operation and its arguments.

pub fn update_application_settings(&self) -> UpdateApplicationSettings<C, M, R>

Constructs a fluent builder for the UpdateApplicationSettings operation.

See UpdateApplicationSettings for more information about the operation and its arguments.

pub fn update_baidu_channel(&self) -> UpdateBaiduChannel<C, M, R>

Constructs a fluent builder for the UpdateBaiduChannel operation.

See UpdateBaiduChannel for more information about the operation and its arguments.

pub fn update_campaign(&self) -> UpdateCampaign<C, M, R>

Constructs a fluent builder for the UpdateCampaign operation.

See UpdateCampaign for more information about the operation and its arguments.

pub fn update_email_channel(&self) -> UpdateEmailChannel<C, M, R>

Constructs a fluent builder for the UpdateEmailChannel operation.

See UpdateEmailChannel for more information about the operation and its arguments.

pub fn update_email_template(&self) -> UpdateEmailTemplate<C, M, R>

Constructs a fluent builder for the UpdateEmailTemplate operation.

See UpdateEmailTemplate for more information about the operation and its arguments.

pub fn update_endpoint(&self) -> UpdateEndpoint<C, M, R>

Constructs a fluent builder for the UpdateEndpoint operation.

See UpdateEndpoint for more information about the operation and its arguments.

pub fn update_endpoints_batch(&self) -> UpdateEndpointsBatch<C, M, R>

Constructs a fluent builder for the UpdateEndpointsBatch operation.

See UpdateEndpointsBatch for more information about the operation and its arguments.

pub fn update_gcm_channel(&self) -> UpdateGcmChannel<C, M, R>

Constructs a fluent builder for the UpdateGcmChannel operation.

See UpdateGcmChannel for more information about the operation and its arguments.

pub fn update_in_app_template(&self) -> UpdateInAppTemplate<C, M, R>

Constructs a fluent builder for the UpdateInAppTemplate operation.

See UpdateInAppTemplate for more information about the operation and its arguments.

pub fn update_journey(&self) -> UpdateJourney<C, M, R>

Constructs a fluent builder for the UpdateJourney operation.

See UpdateJourney for more information about the operation and its arguments.

pub fn update_journey_state(&self) -> UpdateJourneyState<C, M, R>

Constructs a fluent builder for the UpdateJourneyState operation.

See UpdateJourneyState for more information about the operation and its arguments.

pub fn update_push_template(&self) -> UpdatePushTemplate<C, M, R>

Constructs a fluent builder for the UpdatePushTemplate operation.

See UpdatePushTemplate for more information about the operation and its arguments.

pub fn update_recommender_configuration(
&self
) -> UpdateRecommenderConfiguration<C, M, R>

Constructs a fluent builder for the UpdateRecommenderConfiguration operation.

See UpdateRecommenderConfiguration for more information about the operation and its arguments.

pub fn update_segment(&self) -> UpdateSegment<C, M, R>

Constructs a fluent builder for the UpdateSegment operation.

See UpdateSegment for more information about the operation and its arguments.

pub fn update_sms_channel(&self) -> UpdateSmsChannel<C, M, R>

Constructs a fluent builder for the UpdateSmsChannel operation.

See UpdateSmsChannel for more information about the operation and its arguments.

pub fn update_sms_template(&self) -> UpdateSmsTemplate<C, M, R>

Constructs a fluent builder for the UpdateSmsTemplate operation.

See UpdateSmsTemplate for more information about the operation and its arguments.

pub fn update_template_active_version(
&self
) -> UpdateTemplateActiveVersion<C, M, R>

Constructs a fluent builder for the UpdateTemplateActiveVersion operation.

See UpdateTemplateActiveVersion for more information about the operation and its arguments.

pub fn update_voice_channel(&self) -> UpdateVoiceChannel<C, M, R>

Constructs a fluent builder for the UpdateVoiceChannel operation.

See UpdateVoiceChannel for more information about the operation and its arguments.

pub fn update_voice_template(&self) -> UpdateVoiceTemplate<C, M, R>

Constructs a fluent builder for the UpdateVoiceTemplate operation.

See UpdateVoiceTemplate for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.